The InputReplyHander class was renamed to InputReplyHandler; The methods from InputReplyHandler class were summarized; The Id, Type, Value and Direction property values were added into message metadatas.
